move top products querrying from project base to packages
