+++ upgrade-notes/backend_20240628_060247.md
@@ -1,0 +1,11 @@
+#### top products are now in elasticsearch ([#3242](https://github.com/shopsys/shopsys/pull/3242))
+
+-   constructor `Shopsys\FrontendApiBundle\Model\Resolver\Products\PromotedProductsQuery` changed its interface:
+    ```diff
+        public function __construct(
+            protected readonly CurrentCustomerUser $currentCustomerUser,
+    +       protected readonly ProductFacade $productFacade,
+    ```
+-   [features moved](#movement-of-features-from-project-base-to-packages) from project-base to the frontend-api package:
+    -   `PromotedProductsQuery` and all the related logic
+-   see #project-base-diff to update your project
